+++ fabric/src/main/java/cc/cassian/immersiveoverlays/fabric/ImmersiveOverlaysFabricClient.java
@@ -7,6 +7,7 @@ import cc.cassian.immersiveoverlays.overlay.ClockOverlay;
 import cc.cassian.immersiveoverlays.overlay.CompassOverlay;
 import cc.cassian.immersiveoverlays.overlay.OverlayHelpers;
 import net.fabricmc.api.ClientModInitializer;
+import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientLifecycleEvents;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.rendering.v1.HudRenderCallback;
 
@@ -15,12 +16,14 @@ public final class ImmersiveOverlaysFabricClient implements ClientModInitializer
     public void onInitializeClient() {
         // This entrypoint is suitable for setting up client-specific logic, such as rendering.
         ModClient.init();
-        ModLists.loadLists();
         HudRenderCallback.EVENT.register(((matrixStack, tickDelta) -> {
             CompassOverlay.renderGameOverlayEvent(matrixStack);
             ClockOverlay.renderGameOverlayEvent(matrixStack);
         }));
         ClientTickEvents.END_CLIENT_TICK.register(OverlayHelpers::checkInventoryForOverlays);
+        ClientLifecycleEvents.CLIENT_STARTED.register((client -> {
+            ModLists.loadLists();
+        }));
     }
 
 }