+++ forge/src/main/java/cc/cassian/immersiveoverlays/compat/forge/MapAtlasesCompatImpl.java
@@ -1,12 +1,13 @@
 package cc.cassian.immersiveoverlays.compat.forge;
 
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
 import pepjebs.mapatlases.client.MapAtlasesClient;
 import pepjebs.mapatlases.config.MapAtlasesClientConfig;
 
 public class MapAtlasesCompatImpl {
 
     public static boolean showingCoords(Player player) {
-        return MapAtlasesClient.getCurrentActiveAtlas() != null && MapAtlasesClientConfig.drawMinimapCoords.get();
+        return !MapAtlasesClient.getCurrentActiveAtlas().isEmpty() && MapAtlasesClientConfig.drawMinimapCoords.get();
     }
 }