+++ neoforge/src/main/java/cc/cassian/immersiveoverlays/neoforge/ImmersiveOverlaysNeoForgeClient.java
@@ -6,27 +6,35 @@ import cc.cassian.immersiveoverlays.config.neoforge.ModConfigFactory;
 import cc.cassian.immersiveoverlays.overlay.ClockOverlay;
 import cc.cassian.immersiveoverlays.overlay.CompassOverlay;
 import cc.cassian.immersiveoverlays.overlay.OverlayHelpers;
+import cc.cassian.immersiveoverlays.overlay.OverlayLayer;
 import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.Gui;
+import net.minecraft.resources.ResourceLocation;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.ModContainer;
 import net.neoforged.fml.ModLoadingContext;
 import net.neoforged.fml.event.lifecycle.FMLClientSetupEvent;
 import net.neoforged.neoforge.client.event.ClientTickEvent;
+import net.neoforged.neoforge.client.event.RegisterGuiLayersEvent;
 import net.neoforged.neoforge.client.event.RenderGuiLayerEvent;
+import net.neoforged.neoforge.client.gui.GuiLayerManager;
 import net.neoforged.neoforge.client.gui.IConfigScreenFactory;
+import net.neoforged.neoforge.client.gui.VanillaGuiLayers;
 import net.neoforged.neoforge.common.NeoForge;
 
 import static cc.cassian.immersiveoverlays.helpers.ModHelpers.clothConfigInstalled;
 
 public class ImmersiveOverlaysNeoForgeClient {
+    public static final OverlayLayer LAYER = new OverlayLayer();
+
     public static void init(IEventBus eventBus, ModContainer modContainer) {
         // Load config.
         ModClient.init();
         //Register config screen.
         registerModsPage();
         NeoForge.EVENT_BUS.addListener(ImmersiveOverlaysNeoForgeClient::checkInventoryForOverlays);
-        NeoForge.EVENT_BUS.addListener(ImmersiveOverlaysNeoForgeClient::renderGameOverlayEvent);
+        eventBus.addListener(ImmersiveOverlaysNeoForgeClient::renderGameOverlayEvent);
         eventBus.addListener(ImmersiveOverlaysNeoForgeClient::loadComplete);
 
     }
@@ -37,9 +45,8 @@ public class ImmersiveOverlaysNeoForgeClient {
     }
 
     @SubscribeEvent
-    public static void renderGameOverlayEvent(RenderGuiLayerEvent.Post event) {
-        CompassOverlay.renderGameOverlayEvent(event.getGuiGraphics());
-        ClockOverlay.renderGameOverlayEvent(event.getGuiGraphics());
+    public static void renderGameOverlayEvent(RegisterGuiLayersEvent event) {
+        event.registerAboveAll(ResourceLocation.fromNamespaceAndPath(ModClient.MOD_ID, "overlay"), LAYER);
     }
 
     @SubscribeEvent